requirements.txt
Flask==1.1.2
Flask_Script==2.0.6
Flask_SQLAlchemy==2.4.4
Werkzeug==1.0.1
alembic==1.4.3
SQLAlchemy==1.3.19
Flask_Migrate==2.5.3
Jinja2==2.11.2
gunicorn==19.9.0
pyjwt==1.4.2